[refactor] 게시글 상세 모달 - custom hook, import 정리, 댓글등록 컴포넌트 분리
