+++ app/_types/comments/comments.ts
@@ -1,0 +1,4 @@
+export interface AddCommentProps {
+  loggedInUserUid: string;
+  post_id: string;
+}
